Adding two minute grace period on startup
Adding a two minute timer before starting the main program loop. In the event that the script starts throwing exceptions in the main loop - which subsequently run the shutdown server command - this two minute grace period will allow for a user to quickly RDP into the server, and then use services manager to shutdown the Python script so that they can make any necessary changes.